refactored based on code review comments
